move emit_page_viewed_event/1 and its helpers to student's Utils module
